ICOc: Remove uncommon characters from output
The return (⏎) and the rightwards double arrow character (⇒) are not
included in the standard font of Windows PowerShell. The result is that
PowerShell shows a block instead of the characters. Since we assume
that many users of ICOc start the program in this terminal, we use a
textual description instead of the “special” characters.