CLI: Fix sensor channel configuration write back
We now use the special sensor channel number `0` (currently defined as
“no change to channel number”) as default value instead of sensor
channel `1`.

This should not be a problem in connection with the definition in the
CLI interface of ICOc, where `0` is defined as disabling the measurement
channel. Here we would need to enter the sensor channel number to enable
the measurement anyway.

Thank you to @LoydIrrving for reporting the bug.