Storage: Overwrite old data files
While appending data might sound nice in theory – no data loss if you
use the same channel configuration – the problems that come with
appending are not worth it. For example, since the measurement start
date is overwritten, additional measurements are layered on top of the
first measurement instead of starting (at some unspecified time) after
the first measurement.